+++ Runtime/Utils/Math/Extensions/float3_extension.cs
@@ -18,27 +18,19 @@ All rights reserved.
 */
 
 using System.Runtime.CompilerServices;
-using UnityEngine;
 using Unity.Mathematics;
 
 namespace Voxell.Mathx
 {
-  public static class float2_extension
+  public static class float3_extension
   {
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static float2 perpendicular(this float2 vector) => new float2(-vector.y, vector.x);
-
-    // Returns the angle in degrees between /from/ and /to/.
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static float angle(this float2 from, float2 to)
+    public static int LongAxis(this float3 v)
     {
-      // sqrt(a) * sqrt(b) = sqrt(a * b) -- valid for real numbers
-      float denominator = (float)math.sqrt(math.lengthsq(from) * math.lengthsq(to));
-      if (denominator < Vector2.kEpsilonNormalSqrt)
-        return 0F;
-
-      float dot = math.clamp(math.dot(from, to) / denominator, -1F, 1F);
-      return math.acos(dot) * Mathf.Rad2Deg;
+      int i = 0;
+      if (math.abs(v.y) > math.abs(v.x)) i = 1;
+      if (math.abs(v.z) > math.abs(i == 0 ? v.x : v.y)) i = 2;
+      return i;
     }
   }
 }